+++ PocketRoulette/Plugin.cs
@@ -1,10 +1,11 @@
 using BepInEx;
 using BepInEx.Logging;
+using System.IO;
 using PocketRoulette.Patches;
 
 namespace PocketRoulette
 {
-    [BepInPlugin("com.kairo.pocketroulette", "PocketRoulette", "1.1.0")]
+    [BepInPlugin("com.kairo.pocketroulette", "PocketRoulette", "1.2.0")]
     public class Plugin : BaseUnityPlugin
     {
         public static ManualLogSource LogSource = null!;
@@ -13,12 +14,27 @@ namespace PocketRoulette
         private void Awake()
         {
             LogSource = Logger;
-            CachedConfig = Models.PocketRouletteConfig.CreateDefault();
+            ClientConfigManager.Bind(Config, Path.GetDirectoryName(Info.Location));
+            Config.Save();
+            CachedConfig = LoadStartupConfig();
+            ClientConfigManager.EnsureJsonExists(CachedConfig);
 
             new RaidStartPatch().Enable();
             FikaMaybe.Init();
 
             LogSource.LogInfo("pocket roulette loaded");
+        }
+
+        private Models.PocketRouletteConfig LoadStartupConfig()
+        {
+            try
+            {
+                return ConfigLoader.FetchConfig();
+            }
+            catch
+            {
+                return Models.PocketRouletteConfig.CreateDefault();
+            }
         }
     }
 }